change desing pattern
