+++ bindings/vars.py
@@ -1,0 +1,23 @@
+import os
+
+# Set the varible keys
+super = "mod4"
+alt = "mod1"
+print_screen = "Print"
+
+
+# set the varible locations
+home = os.path.expanduser("~")
+app_menu = f"{home}/.config/qtile/scripts/rofi/dmenu_drun"
+shell_menu = f"{home}/.config/qtile/scripts/rofi/dmenu_run"
+win_selector = f"{home}/.config/qtile/scripts/rofi/dmenu_window"
+screenshot = f"{home}/.config/qtile/scripts/screenshot"
+
+# set the varible apps
+term = "alacritty"
+clipboard = (
+    'rofi -modi "clipboard:greenclip print" -show clipboard -run-command "{cmd}"'
+)
+browser = "firefox"
+file_manager = "pcmanfm-qt"
+qt5_config = "qt5ct"
