+++ groups/names.py
@@ -1,0 +1,11 @@
+# Define groups names
+def init_groups():
+    group_names = [
+        ("Web", {"layout": "Bsp"}),
+        ("Code", {"layout": "Bsp"}),
+        ("Music", {"layout": "Bsp"}),
+        ("Game", {"layout": "Bsp"}),
+        ("Files", {"layout": "Bsp"}),
+        ("Others", {"layout": "Bsp"}),
+    ]
+    return group_names
