+++ scripts/rofi/dmenu_drun
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+rofi -show drun -theme ~/.config/qtile/themes/rofi/dracula_demnu_style.rasi
