+++ scripts/rofi/dmenu_run
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+rofi -show run -theme ~/.config/qtile/themes/rofi/darcula_demnu_style.rasi
+
