+++ scripts/rofi/dmenu_window
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+rofi -show window -theme ~/.config/qtile/themes/rofi/dracula_center_style.rasi
