+++ themes/rofi/dracula_demnu_style.rasi
@@ -1,0 +1,57 @@
+* {
+    background-color:           #282a36;
+    border-color:               #f8f8f2;
+    text-color:                 #f8f8f2;
+    height:                     20px;
+    font : "FantasqueSansMono Nerd Font Mono 9" ;
+    prompt-font:                "FantasqueSansMono Nerd Font Mono 10";
+    prompt-background:          #ff79c6;
+    prompt-foreground:          #f8f8f2;
+    prompt-padding:             2px;
+    selected-normal-background: #8be9fd;
+    selected-normal-foreground: #44475a;
+}
+#window {
+    anchor: north;
+    location: north;
+    width: 100%;
+    padding: 0px;
+    children: [ horibox ];
+}
+#horibox {
+    orientation: horizontal;
+    children: [ prompt, entry, listview ];
+}
+#prompt {
+    padding:          @prompt-padding;
+    background-color: @prompt-background;
+    text-color:       @prompt-foreground;
+    font:             @prompt-font;
+}
+#listview {
+    layout: horizontal;
+    lines: 100;
+}
+#entry {
+    padding: 2px;
+    expand: false;
+    width: 10em;
+}
+#element {
+    padding: 2px 8px;
+}
+#element selected {
+    background-color: @selected-normal-background;
+    text-color:       @selected-normal-foreground;
+}
+
+element-text {
+    background-color: inherit;
+    text-color:       inherit;
+}
+
+
+element-icon {
+    background-color: inherit;
+}
+
